+++ examples/shardingsphere-jdbc-example/transaction-example/transaction-2pc-xa-narayana-raw-jdbc-example/src/main/java/org/apache/shardingsphere/example/transaction/xa/narayana/raw/jdbc/TransactionXaNarayanaRawExample.java
@@ -26,8 +26,8 @@ import java.io.File;
 import java.io.IOException;
 import java.sql.SQLException;
 
-public final class ExampleMain {
-    
+public final class TransactionXaNarayanaRawExample {
+
     public static void main(final String[] args) throws SQLException, IOException {
         DataSource dataSource = YamlShardingSphereDataSourceFactory.createDataSource(getFile("/META-INF/sharding-databases-tables.yaml"));
         ExampleExecuteTemplate.run(getExampleService(dataSource));
@@ -36,7 +36,7 @@ public final class ExampleMain {
     }
     
     private static File getFile(final String fileName) {
-        return new File(ExampleMain.class.getResource(fileName).getFile());
+        return new File(TransactionXaNarayanaRawExample.class.getResource(fileName).getFile());
     }
     
     private static ExampleService getExampleService(final DataSource dataSource) {
